Convert deprecated EtcdDiscoveryURL to new DefaultEtcdClusterToken (#69)
Gracefully handle existing etcd discovery url in cluster.json.

If --etcd-discovery and the old EtcdDiscoveryURL are the same discovery -> token is stored and EtcdDiscoveryURL is removed
If the internal etcd discovery is used (which is the default) then the token is stored in etcd and the EtcdDiscoveryURL is removed. Note: the machine/member data is not transferred. Also existing machines will still have the old discovery URL in their cloud-config.
If --etcd-discovery and the old EtcdDiscoveryURLdo not match startup is refused and needs to be fixed manually.